Bug fixes on coworkers.blade

+++ app/Http/Controllers/ToDoController.php
@@ -16,7 +16,8 @@ class ToDoController extends Controller
         if(Auth::user()->is_admin) {
             $coworkers = Invitation::where('admin_id', Auth::user()->id)->where('accepted', 1)->get();
             $invitations = Invitation::where('admin_id', Auth::user()->id)->where('accepted', 0)->get();
-            $tasks = Task::where('user_id', Auth::user()->id)->orWhere('admin_id', Auth::user()->admin_id)->orderBy('created_at', 'DESC')->paginate(4);
+            $tasks = Task::where('user_id', Auth::user()->id)->orWhere('admin_id', Auth::user()->id)->orderBy('created_at', 'DESC')->paginate(4);
+            return view('index', compact('tasks', 'coworkers', 'invitations'));
         } else {
             //use where method on Task MODEL to retrieve all tasks from current user,
             //so they are available on the view
@@ -34,7 +35,18 @@ class ToDoController extends Controller
         if($request->input('task')) {
             $task = new Task;
             $task->content = $request->input('task');
-            Auth::user()->tasks()->save($task);
+
+            if(Auth::user()->is_admin) {
+                if($request->input('assignTo') == Auth::user()->id) {
+                    Auth::user()->tasks()->save($task);
+                } elseif ($request->input('assignTo') != null) {
+                    $task->user_id = $request->input('assignTo');
+                    $task->admin_id = Auth::user()-id;
+                    $task->save();
+                }
+            } else {
+                Auth::user()->tasks()->save($task);
+            }
         }
 
         return redirect()->back();
@@ -43,7 +55,16 @@ class ToDoController extends Controller
     public function edit($id) {
 
         $task = Task::find($id);
-        return view('edit', compact('task'));
+        if(Auth::user()->is_admin) {
+            $coworkers = Invitation::where('admin_id', Auth::user()->id)->where('accepted', 1)->get();
+            $invitations = Invitation::where('admin_id', Auth::user()->id)->where('accepted', 0)->get();
+        } else {
+            $coworkers = [];
+            $invitations = [];
+        }
+
+        return view('edit', compact('task', 'coworkers', 'invitations'));
+        // return view('edit', ['task'=>$task, 'coworkers'=>$coworkers, 'invitations'=>$invitations]);
     }
 
     public function update($id, Request $request) {
@@ -51,7 +72,20 @@ class ToDoController extends Controller
         if($request->input('task')){
             $task = Task::find($id);
             $task->content = $request->input('task');
-            $task->save();
+
+            if(Auth::user()->is_admin) {
+                if($request->input('assignTo') == Auth::user()->id) {
+                    Auth::user()->tasks()->save($task);
+                } elseif ($request->input('assignTo') != null) {
+                    $task->user_id = $request->input('assignTo');
+                    $task->admin_id = Auth::user()-id;
+                    $task->save();
+                }
+            } else {
+                if($this->_authorize($task->user_id))
+                $task->save();
+            }
+
         }
         return redirect('/');
     }
@@ -59,6 +93,14 @@ class ToDoController extends Controller
     public function delete($id) {
 
         $task = Task::find($id);
+
+        if(!Auth::user()->is_admin) {
+            if(!$this->_authorize($task->user_id)){
+                return redirect()->back();
+                exit();
+            }
+        }
+
         $task->delete();
         return redirect()->back();
     }
@@ -67,6 +109,7 @@ class ToDoController extends Controller
 
         $task = Task::find($id);
         $task->status = !$task->status;
+        if($this->_authorize($task->user_id))
         $task->save();
         return redirect()->back();
     }
@@ -84,18 +127,33 @@ class ToDoController extends Controller
     }
 
     public function acceptInvitation(Request $request) {
-        $invitation = Invitation::findO($id);
+        $invitation = Invitation::find($id);
         $invitation->accepted = true;
+        if($this->_authorize($invitation->admin_id))
         $invitation->save();
 
         return redirect()->back();
     }
 
     public function denyInvitation(Request $request) {
-        $invitation = Invitation::findO($id);
+        $invitation = Invitation::find($id);
+        if($this->_authorize($invitation->admin_id))
         $invitation->delete();
 
         return redirect()->back();
 
+    }
+
+    public function deleteWorker($id) {
+        $invitation = Invitation::find($id);
+        if($this->_authorize($invitation->admin_id))
+        $invitation->delete();
+
+        return redirect()->back();
+
+    }
+
+    protected function _authorize($id) {
+        return Auth::user()->id === $id ? true : false;
     }
 }
